requirements.txt
pandas
PyPDF2
chromadb
chainlit
python-dotenv
google-generativeai==0.3.2
langchain
